moved app.js to src folder
